A BASIC WEATHER APP
I created this basic weather app that fetches data from the openweathermap API (more details in readme file) and shows the weather of a city that you enter in the search bar. Tried to keep the design minimalistic. Used HTML,CSS & JS. Will try to improve this version or maybe try and create a better version with more functionalities and use a UI components library for an even better UI.